backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.63.1
pydantic==2.9.2
pydantic-settings==2.6.0
pandas==2.2.3
numpy==1.26.4
faiss-cpu==1.8.0.post1
openpyxl==3.1.5
python-dotenv==1.0.1
